Refactor common code into AbstractRenditionIntegrationTest in order to get context set correctly for RenditionTest.
Also added *.url values (to use docker images) to the test a-g.p so we don't have to keep adding them to tests in Intellij or Bamboo.
