+++ src/test/resources/alfresco-global.properties
@@ -44,3 +44,7 @@ identity-service.ignore-oauth-query-parameter=true
 identity-service.credentials.secret=11111
 identity-service.credentials.provider=secret
 
+alfresco-pdf-renderer.url=http://localhost:8090/ 
+jodconverter.url=http://localhost:8092/ 
+img.url=http://localhost:8091/ 
+tika.url=http://localhost:8093/
